Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/referrer-policy/origin/attr-referrer/same-origin/http-http/script-tag/keep-origin-redirect/generic.http-expected.txt


FAIL The referrer URL is origin when a
                                 document served over http requires an http
                                 sub-resource via script-tag using the attr-referrer
                                 delivery method with keep-origin-redirect and when
                                 the target request is same-origin. assert_equals: Reported Referrer URL is 'origin'. expected "http://localhost:8800/" but got "http://localhost:8800/referrer-policy/origin/attr-referrer/same-origin/http-http/script-tag/keep-origin-redirect/generic.http.html"
